backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
redis==5.2.1
transformers==4.46.3